subcomposer is now a module.  You can still run it from the command
subcomposer is now a module.  You can still run it from the command
line.  If you want it to work as a module, you need to call the
considersendupdate() method with the period of the after loop that it
should use.

darcs-hash:20030702102159-13ce2-e27e340c319d48768684a83ea00ee65d89b3beb5.gz